+++ wasmbinding/exec_perp.go
@@ -1,11 +1,6 @@
 package wasmbinding
 
 import (
-	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/NibiruChain/nibiru/wasmbinding/bindings"
-	"github.com/NibiruChain/nibiru/x/common/asset"
 	perpv2keeper "github.com/NibiruChain/nibiru/x/perp/v2/keeper"
 	perpv2types "github.com/NibiruChain/nibiru/x/perp/v2/types"
 )
@@ -16,20 +11,4 @@ type ExecutorPerp struct {
 
 func (exec *ExecutorPerp) MsgServer() perpv2types.MsgServer {
 	return perpv2keeper.NewMsgServerImpl(exec.PerpV2)
-}
-
-// TODO: rename to CloseMarket
-func (exec *ExecutorPerp) SetMarketEnabled(
-	cwMsg *bindings.SetMarketEnabled, ctx sdk.Context,
-) (err error) {
-	if cwMsg == nil {
-		return wasmvmtypes.InvalidRequest{Err: "null msg"}
-	}
-
-	pair, err := asset.TryNewPair(cwMsg.Pair)
-	if err != nil {
-		return err
-	}
-
-	return exec.PerpV2.Sudo().CloseMarket(ctx, pair)
 }